function_app/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pydantic==2.8.2
pandas==2.2.2
jsonschema==4.23.0
pypdf==4.2.0
PyMuPDF==1.24.4
pillow==10.3.0
openai==1.30.3
tabulate==0.9.0
python-dotenv==1.0.1
azure-identity==1.17.1
azure-functions==1.20.0
azure-ai-documentintelligence==1.0.0b2
azure-ai-formrecognizer==3.3
# Install haystack-ai fork with multimodal support - see https://github.com/deepset-ai/haystack/pull/7943/files
https://github.com/CarlosFerLo/haystack/archive/3c2a913688e94001339d5dc3461da2862a28dac9.tar.gz
pytest==8.3.2